Move lock.rs into its own module
